Comments working for both services and crowdfundings.
